Check if substvars exists before operating on it

+++ ci-tooling/lib/ci/generate_langpack_packaging.rb
@@ -30,9 +30,11 @@ module CI
         File.write(file, subbed)
       end
 
-      substvars = File.open('debian/substvars').read
-      substvars.gsub!(/aaaADDITIONALDEPSbbb/, '')
-      File.write('debian/substvars', substvars)
+      if File.exist?('debian/substvars')
+        substvars = File.open('debian/substvars').read
+        substvars.gsub!(/aaaADDITIONALDEPSbbb/, '')
+        File.write('debian/substvars', substvars)
+      end
     end
   end
 end
